+++ manager/pkg/status/conflator/conflation_manager.go
@@ -7,7 +7,7 @@ import (
 	"github.com/go-logr/logr"
 	ctrl "sigs.k8s.io/controller-runtime"
 
-	"github.com/stolostron/multicluster-global-hub/manager/pkg/statussyncer/conflator/metadata"
+	"github.com/stolostron/multicluster-global-hub/manager/pkg/status/conflator/metadata"
 	"github.com/stolostron/multicluster-global-hub/pkg/statistics"
 	"github.com/stolostron/multicluster-global-hub/pkg/transport/consumer"
 )
