--- manager/pkg/status/syncers/placements_handler.go
@@ -1,19 +1,0 @@
-package dbsyncer
-
-import (
-	"fmt"
-
-	clustersv1beta1 "open-cluster-management.io/api/cluster/v1beta1"
-
-	"github.com/stolostron/multicluster-global-hub/manager/pkg/statussyncer/conflator"
-	"github.com/stolostron/multicluster-global-hub/pkg/database"
-	"github.com/stolostron/multicluster-global-hub/pkg/enum"
-)
-
-func NewPlacementHandler() conflator.Handler {
-	return NewGenericHandler[*clustersv1beta1.Placement](
-		string(enum.PlacementSpecType),
-		conflator.PlacementPriority,
-		enum.CompleteStateMode,
-		fmt.Sprintf("%s.%s", database.StatusSchema, database.PlacementsTableName))
-}
