+++ test/integration/manager/api/api_test.go
@@ -18,8 +18,8 @@ import (
 	. "github.com/onsi/gomega"
 	"gorm.io/gorm"
 
-	"github.com/stolostron/multicluster-global-hub/manager/pkg/nonk8sapi"
-	"github.com/stolostron/multicluster-global-hub/manager/pkg/nonk8sapi/util"
+	"github.com/stolostron/multicluster-global-hub/manager/pkg/restapis"
+	"github.com/stolostron/multicluster-global-hub/manager/pkg/restapis/util"
 	"github.com/stolostron/multicluster-global-hub/pkg/database"
 	"github.com/stolostron/multicluster-global-hub/pkg/database/models"
 )
